rag-mcp-example/base/pyproject.toml
[project]
name = "django-rag-api"
version = "1.0.0"
description = "Django Documentation RAG API"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "pydantic>=2.5.0",
    "anthropic>=0.39.0",
    "mcp>=1.0.0",
    "pypdf>=3.0.0",
    "numpy>=1.24.0",
    "pydantic-settings>=2.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
